projects/simple_chatbot/requirements.txt
openai==1.51.2
streamlit==1.39.0
fastapi==0.115.0
python-dotenv==1.0.1
requests==2.32.3